Task-8, Refactoring to make the code a bit cleaner
Found that some of the methods in TimeUtil classes has public access specifier which is not required since its scope is only within TimeUtil class.

Changing TimeUtil to enum make the code more cleaner and short  when calling the utility method.

The right place of Format enum is inside TimeUtil as nested enum

Ammended unit tests

Bumped the version in pom.xml